backend/requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tinycss2==1.0.2
webencodings==0.5.1
Werkzeug==1.0.1
